+++ @katesowles/lab-katesowles/test/greet.test.js
@@ -4,10 +4,11 @@ var greet = require("../greet");
 describe ("greetings", () => {
     it("greets properly", function() {
         const greeting = greet("kate");
-        assert.ok( /hello kate, how is your day going?/.test(greeting), `result was "${greeting}"`);
+        assert.ok( /hello kate, how is your (Mon|Tues|Wednes|Thurs|Fri|Satur|Sun)day going?/.test(greeting), `result was "${greeting}"`);
     });
     it("no name = no one", function() {
         const greeting = greet();
-        assert.ok( /hello no one, how is your day going?/.test(greeting), `result was "${greeting}"`);
+        assert( /hello no one, how is your (Mon|Tues|Wednes|Thurs|Fri|Satur|Sun)day going?/.test(greeting), `result was "${greeting}"`);
     });
+
 });